requirements.txt
asgiref==3.7.2
distlib==0.3.7
Django==4.2.4
filelock==3.12.3
gunicorn==21.2.0
packaging==23.1
platformdirs==3.10.0
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.24.4
whitenoise==6.5.0
django-filter==23.5.0
Pillow
